refactor(cli): split config command implementation
